+++ modules/sdk-core/src/bitgo/wallet/wallet.ts
@@ -99,9 +99,9 @@ import {
   BulkWalletShareKeychain,
   ManageUnspentReservationOptions,
   SignAndSendTxRequestOptions,
+  FundForwarderParams,
 } from './iWallet';
 import { GoStakingWallet, StakingWallet } from '../staking';
-import { Lightning } from '../lightning/custodial';
 import EddsaUtils from '../utils/tss/eddsa';
 import { EcdsaMPCv2Utils, EcdsaUtils } from '../utils/tss/ecdsa';
 import { getTxRequest } from '../tss';
@@ -3101,16 +3101,6 @@ export class Wallet implements IWallet {
     };
   }
 
-  /**
-   * Create lightning for btc/tbtc from this wallet
-   */
-  public lightning(): Lightning {
-    if (!this.baseCoin.supportsLightning()) {
-      throw new Error(`Lightning not supported for ${this.coin()}`);
-    }
-    return new Lightning(this.bitgo, this);
-  }
-
   /* MARK: TSS Helpers */
 
   /**
@@ -3636,6 +3626,22 @@ export class Wallet implements IWallet {
       throw new Error('forwarder address required');
     }
     const url = this.url('/fundForwarder');
+    this._wallet = await this.bitgo.post(url).send(params).result();
+    return this._wallet;
+  }
+
+  /**
+   * Send funds from a fee address to a forwarder.
+   *
+   * @param {Object} params - parameters object
+   * @param {List} params.forwarders - list of fund forwarder options
+   * @returns {*}
+   */
+  public async fundForwarders(params: FundForwarderParams): Promise<any> {
+    if (_.isUndefined(params.forwarders) || params.forwarders.length == 0) {
+      throw new Error('fund forwarder options required');
+    }
+    const url = this.url('/fundforwarders');
     this._wallet = await this.bitgo.post(url).send(params).result();
     return this._wallet;
   }