+++ modules/abstract-lightning/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-lightning@4.2.0...@bitgo/abstract-lightning@4.2.1) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/abstract-lightning
+
 # [4.2.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-lightning@4.1.1...@bitgo/abstract-lightning@4.2.0) (2025-03-28)
 
 ### Features