+++ modules/abstract-utxo/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.19.4](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-utxo@9.19.3...@bitgo/abstract-utxo@9.19.4) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/abstract-utxo
+
 ## [9.19.3](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-utxo@9.19.2...@bitgo/abstract-utxo@9.19.3) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/abstract-utxo