+++ modules/sdk-api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.61.4](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-api@1.61.3...@bitgo/sdk-api@1.61.4) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-api
+
 ## [1.61.3](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-api@1.61.2...@bitgo/sdk-api@1.61.3) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-api