+++ modules/sdk-coin-baby/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.3](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-baby@1.2.2...@bitgo/sdk-coin-baby@1.2.3) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-baby
+
 ## [1.2.2](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-baby@1.2.1...@bitgo/sdk-coin-baby@1.2.2) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-baby