+++ modules/sdk-coin-bch/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.24](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-bch@2.1.23...@bitgo/sdk-coin-bch@2.1.24) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-bch
+
 ## [2.1.23](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-bch@2.1.22...@bitgo/sdk-coin-bch@2.1.23) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-bch