+++ modules/sdk-coin-bsc/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [22.2.27](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-bsc@22.2.26...@bitgo/sdk-coin-bsc@22.2.27) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-bsc
+
 ## [22.2.26](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-bsc@22.2.25...@bitgo/sdk-coin-bsc@22.2.26) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-bsc