+++ modules/sdk-coin-etc/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.41](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-etc@2.2.40...@bitgo/sdk-coin-etc@2.2.41) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-etc
+
 ## [2.2.40](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-etc@2.2.39...@bitgo/sdk-coin-etc@2.2.40) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-etc