+++ modules/sdk-coin-eth/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [24.2.63](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eth@24.2.62...@bitgo/sdk-coin-eth@24.2.63) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-eth
+
 ## [24.2.62](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eth@24.2.61...@bitgo/sdk-coin-eth@24.2.62) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-eth