+++ modules/sdk-coin-ethlike/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.9](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ethlike@1.2.8...@bitgo/sdk-coin-ethlike@1.2.9) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-ethlike
+
 ## [1.2.8](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ethlike@1.2.7...@bitgo/sdk-coin-ethlike@1.2.8) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-ethlike