+++ modules/sdk-coin-hash/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.43](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-hash@3.0.42...@bitgo/sdk-coin-hash@3.0.43) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-hash
+
 ## [3.0.42](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-hash@3.0.41...@bitgo/sdk-coin-hash@3.0.42) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-hash