+++ modules/sdk-coin-hbar/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.74](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-hbar@2.0.73...@bitgo/sdk-coin-hbar@2.0.74) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-hbar
+
 ## [2.0.73](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-hbar@2.0.72...@bitgo/sdk-coin-hbar@2.0.73) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-hbar