+++ modules/sdk-coin-lnbtc/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.5](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-lnbtc@1.2.4...@bitgo/sdk-coin-lnbtc@1.2.5) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-lnbtc
+
 ## [1.2.4](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-lnbtc@1.2.3...@bitgo/sdk-coin-lnbtc@1.2.4) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-lnbtc