+++ modules/sdk-coin-opeth/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [18.1.59](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-opeth@18.1.58...@bitgo/sdk-coin-opeth@18.1.59) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-opeth
+
 ## [18.1.58](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-opeth@18.1.57...@bitgo/sdk-coin-opeth@18.1.58) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-opeth