+++ modules/sdk-coin-sol/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.13](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-sol@4.8.12...@bitgo/sdk-coin-sol@4.8.13) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-sol
+
 ## [4.8.12](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-sol@4.8.11...@bitgo/sdk-coin-sol@4.8.12) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-sol