+++ modules/sdk-coin-stx/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.3](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-stx@3.5.2...@bitgo/sdk-coin-stx@3.5.3) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-stx
+
 ## [3.5.2](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-stx@3.5.1...@bitgo/sdk-coin-stx@3.5.2) (2025-03-28)
 
 ### Bug Fixes