+++ modules/sdk-coin-ton/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.31](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ton@3.4.30...@bitgo/sdk-coin-ton@3.4.31) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-ton
+
 ## [3.4.30](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ton@3.4.29...@bitgo/sdk-coin-ton@3.4.30) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-ton