+++ modules/sdk-unified-wallet/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.0.74](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-unified-wallet@11.0.73...@bitgo/sdk-unified-wallet@11.0.74) (2025-04-02)
+
+**Note:** Version bump only for package @bitgo/sdk-unified-wallet
+
 ## [11.0.73](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-unified-wallet@11.0.72...@bitgo/sdk-unified-wallet@11.0.73) (2025-03-28)
 
 **Note:** Version bump only for package @bitgo/sdk-unified-wallet